fix 'ascii' codec can't encode character X in position... error
This error happens if the webpage contains non-ascii characters. Let's
decode from a more flexible 'utf-8' since many webpage are nowadays
encoded in utf-8. A more complete solution would look at the different
locations (http-header, html-header) which specify the encoding.

To avoid errors when we receive sloppy formatted utf-8 (utf-8 which also
contains some invalid characters) we use the replace error handler for
decode.

To stay compatible with the cache we encode back to utf-8 when writing
the cache.

We also need to encode when sending to stdout. Since when run under crontab the
encoding of the terminal might not be set correctly. This is a general problem
when piping output from python into other processes, because python uses what
the outside defines.